pyproject.toml
[project]
name = "api-scandish"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "langchain>=0.3.23",
    "langchain-together>=0.3.0",
    "numpy>=2.2.5",
    "pillow>=11.2.1",
    "python-multipart>=0.0.20",
    "supabase>=2.15.0",
    "torchmetrics>=1.7.1",
]
